Added docker files and fix issue with query build
