+++ data_sync_microservice/requirements.txt
@@ -1,0 +1,22 @@
+APScheduler==3.10.4
+black==24.4.2
+certifi==2024.6.2
+charset-normalizer==3.3.2
+click==8.1.7
+greenlet==3.0.3
+idna==3.7
+mypy-extensions==1.0.0
+packaging==24.1
+pathspec==0.12.1
+platformdirs==4.2.2
+psycopg2==2.9.9
+py4j==0.10.9.7
+pyspark==3.5.1
+pytz==2024.1
+requests==2.32.3
+six==1.16.0
+SQLAlchemy==2.0.30
+tomli==2.0.1
+typing_extensions==4.12.2
+tzlocal==5.2
+urllib3==2.2.1
